requirements.txt
# put your requirements here
pyqtgraph
PyQt5
numpy
pyserial
